+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [next]
 
+## [2.9.0] - 2025-09-07
+
 ### Added
 
 - Add `IntellijIdea` platform type with corresponding artifact coordinates. JetBrains/intellij-platform-gradle-plugin#2021
@@ -1538,7 +1540,8 @@ The `2.0.0` release is completely rewritten. Please see [documentation page](htt
 
 - Support for attaching IntelliJ sources in IDEA
 
-[next]: https://github.com/JetBrains/intellij-platform-gradle-plugin/compare/v2.8.0...HEAD
+[next]: https://github.com/JetBrains/intellij-platform-gradle-plugin/compare/v2.9.0...HEAD
+[2.9.0]: https://github.com/JetBrains/intellij-platform-gradle-plugin/compare/v2.8.0...v2.9.0
 [2.8.0]: https://github.com/JetBrains/intellij-platform-gradle-plugin/compare/v2.7.2...v2.8.0
 [2.7.2]: https://github.com/JetBrains/intellij-platform-gradle-plugin/compare/v2.7.1...v2.7.2
 [2.7.1]: https://github.com/JetBrains/intellij-platform-gradle-plugin/compare/v2.7.0...v2.7.1